third_party/WebKit/LayoutTests/fast/dom/anchor-backslash-expected.txt

This tests \ characters being changed to / in anchor tags.

If it works you should see five messages below that have \ or / characters as specified.

1 2 3 4 5

paths like this should turn all \ characters into /: "/a/b/c/"
query strings like this one should still have \ characters: "?path=\\a\b\c\"
query strings like this one should still have \ characters even if the URL is only a query: "?path=\\myshare\myfolder\myfile\"
anchor strings like this one should still have \ characters: "#path=\\a\b\c\"
anchor strings like this one should still have \ characters even if the URL is only an anchor: "#path=\\myshare\myfolder\myfile\"
